refactor(skill): simplify EmptyTip component and update SkillSelectModal logic
