+++ projects/app/src/pageComponents/app/detail/Edit/FormComponent/ToolSelector/hooks/useSkillSelectData.ts
@@ -5,7 +5,7 @@ import type { ListSkillsResponse } from '@fastgpt/global/core/ai/skill/api';
 import type { ParentIdType } from '@fastgpt/global/common/parentFolder/type';
 
 export type SkillSelectItemType = ListSkillsResponse['list'][number];
-export type SkillSelectNavItemType = { id: string; name: string; hasWritePer: boolean };
+export type SkillSelectNavItemType = { id: string; name: string };
 
 /**
  * 维护 Skill 选择弹窗的数据源。
@@ -22,9 +22,6 @@ export const useSkillSelectData = () => {
   // - parentId（根目录为 null，ParentIdType）是文件夹模型 / 创建·导入弹窗的入参约定。
   const fetchParentId = navStack.length > 0 ? navStack[navStack.length - 1].id : '';
   const parentId: ParentIdType = fetchParentId || null;
-  // 当前所在文件夹的写权限（根目录为 null，由调用方回退到团队级创建权限）。
-  const currentFolderHasWritePer =
-    navStack.length > 0 ? navStack[navStack.length - 1].hasWritePer : null;
 
   const {
     data: skillList = [],
@@ -48,10 +45,7 @@ export const useSkillSelectData = () => {
   );
 
   const onEnterFolder = useCallback((item: SkillSelectItemType) => {
-    setNavStack((prev) => [
-      ...prev,
-      { id: item._id, name: item.name, hasWritePer: item.permission.hasWritePer }
-    ]);
+    setNavStack((prev) => [...prev, { id: item._id, name: item.name }]);
     setSearchKey('');
   }, []);
 
@@ -79,7 +73,6 @@ export const useSkillSelectData = () => {
     setSearchKey,
     paths,
     parentId,
-    currentFolderHasWritePer,
     refreshSkillList,
     onEnterFolder,
     onUpdateParentId
